# part3/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Phonebook API",
    "description": "Phonebook API",
    "version": "1.0.0"
  },
  "host": "https://phonebook-io5n.onrender.com/",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/persons": {
      "description": "Get all persons in the phonebook",
      "get": {
        "description": "Get all persons in the phonebook",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Add a new person to the phonebook",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "xxxx",
                  "required": true
                },
                "number": {
                  "type": "string",
                  "example": "040-123456",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/persons/{id}": {
      "get": {
        "description": "Get person by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Delete person by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "description": "Update person's phone number by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "number": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/info": {
      "get": {
        "description": "Get information about the phonebook",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}